有給付与の編集機能追加

+++ db/migrate/20210722000306_create_remaining_holidays.rb
@@ -1,15 +1,16 @@
 class CreateRemainingHolidays < ActiveRecord::Migration[6.0]
   def change
     create_table :remaining_holidays do |t|
-      t.integer    :last_year_days                        , null: false
+      t.float      :last_year_days                        , null: false
       t.integer    :last_year_time                        , null: false
-      t.integer    :this_year_days                        , null: false
+      t.float      :this_year_days                        , null: false
       t.float      :total_days                            , null: false
       t.integer    :total_time                            , null: false
       t.string     :working_hours                         , null: false
       t.references :user, foreign_key: true               , null: false
       t.float      :changeable_days                       
-      t.integer    :changeable_time       
+      t.integer    :changeable_time    
+      # 申請により、引かれた残りの時間をtime_limitと定義している   
       t.integer    :time_limit
       t.timestamps
